bugfix / CPLP-2111 / idp cleanup leftovers (#322)
* idp cleanup leftovers

* linter fixes

* adjust overlay closing time
