+++ cx-portal/src/components/pages/IDPManagement/IDPListItem.tsx
@@ -20,17 +20,13 @@
 
 import { IdentityProvider } from 'features/admin/idpApiSlice'
 import { useDispatch } from 'react-redux'
-import {
-  DropdownMenu,
-  Image,
-  MenuItem,
-  Tooltips,
-} from 'cx-portal-shared-components'
+import { DropdownMenu, MenuItem } from 'cx-portal-shared-components'
 import { show } from 'features/control/overlay/actions'
 import { OVERLAYS } from 'types/Constants'
 import { useTranslation } from 'react-i18next'
 import './style.scss'
 import { useState } from 'react'
+import IDPStateProgress from './IDPStateProgress'
 
 export default function IDPListItem({
   idp,
@@ -45,7 +41,6 @@ export default function IDPListItem({
   const dispatch = useDispatch()
   const [open, setOpen] = useState<boolean>(false)
   const toggle = () => setOpen(!open)
-  const configured = idp.oidc?.clientId ? true : false
 
   const doConfirmDelete = (e: React.MouseEvent<HTMLElement, MouseEvent>) => {
     e.stopPropagation()
@@ -110,26 +105,6 @@ export default function IDPListItem({
     )
   }
 
-  const getStateConjunction = (enabled: boolean) =>
-    enabled ? ' &' : `, ${t('field.not')}`
-
-  const getStateText = (configured: boolean, enabled: boolean) =>
-    configured
-      ? `${t('field.configured')}${getStateConjunction(enabled)} ${t(
-          'field.enabled'
-        )}`
-      : `${t('field.not')} ${t('field.configured')}`
-
-  const getStateImage = (configured: boolean, enabled: boolean) =>
-    `<svg viewBox="-6 -6 112 36" xmlns="http://www.w3.org/2000/svg"><g stroke="gray" stroke-width="2">
-    <path d="M 12 12 H 80"/>
-    <circle fill="gray" cx="12" cy="12" r="11"/>
-    <circle fill="${configured ? 'gray' : 'white'}" cx="50" cy="12" r="11"/>
-    <circle fill="${enabled ? 'gray' : 'white'}" cx="88" cy="12" r="11"/>
-    </g></svg>`
-
-  const stateMessage = getStateText(configured, idp.enabled)
-
   return (
     <>
       <div className="idp-list-item">
@@ -137,18 +112,7 @@ export default function IDPListItem({
         <span className="name">{idp.displayName || '-'}</span>
         <span className="alias">{idp.alias}</span>
         <span className="state">
-          <Tooltips tooltipPlacement="left" tooltipText={stateMessage}>
-            <div>
-              <Image
-                alt={stateMessage}
-                style={{ width: '112px', height: '40px' }}
-                src={`data:image/svg+xml;utf8,${getStateImage(
-                  configured,
-                  idp.enabled
-                )}`}
-              />
-            </div>
-          </Tooltips>
+          <IDPStateProgress idp={idp} />
         </span>
         <span className={'action menu'}>{renderMenu()}</span>
       </div>
